requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
livekit==1.0.11
livekit-agents==1.1.4
livekit-plugins-openai==1.1.4
livekit-plugins-elevenlabs==1.1.0
motor==3.3.2
pymongo==4.6.0
beanie==1.24.0
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.5
python-dotenv==1.0.0
httpx==0.25.2